Update readme-ascension.html
Remove misleading remark. The goal is to have but one Ascension version, right?

Additional compatibility information
For BG2EE and EET
    Longer Road https://github.com/InfinityMods/LongerRoadEE
    Turnabout https://github.com/RoxanneSHS/TurnaboutEE
    have been updated to work with Ascension 2.x
(There are no known maintainers for the non-EE versions, although the required changes might be relatively simple - testing them of course is not simple.)

Iylios and Sandrah in BG2EE and EET are compatible.
I am not sure about Kelsey and Weimer's Solaufein.